rename the bot
