+++ bots/synchrotron/bot.definition.ts
