+++ bots/synchrotron/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@bp-bots/file-sync-bot",
+  "name": "@bp-bots/synchrotron",
   "scripts": {
     "postinstall": "genenv -o ./.genenv/index.ts -e FILESYNC_DROPBOX_CLIENT_ID -e FILESYNC_DROPBOX_CLIENT_SECRET -e FILESYNC_DROPBOX_ACCESS_TOKEN",
     "check:type": "tsc --noEmit",
