+++ bots/synchrotron/tsconfig.json
